Add GetTokenLink component for external links in PoolCardTri
