+++ src/components/earn/FarmsPortfolio/GetTokenLink/GetTokenLink.constants.ts
@@ -1,0 +1,9 @@
+import { STNEAR, AUSDO } from '../../../../constants/tokens'
+
+export const getTokenList = [
+  { token: STNEAR, link: 'https://metapool.app/dapp/mainnet/metapool-aurora/' },
+  {
+    token: AUSDO,
+    link: 'https://v3.oin.finance/'
+  }
+]
